+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##  [Unrelease]
+### Added
+- Non-interactive mode for PDB procedure
+
 ## [v0.3.0] - 2020-03-20
 ### Added
 - Settings remembered between uses